fix: divide CategoryManageView from AdminView
